...And import math

+++ parcels/particleset.py
@@ -7,6 +7,7 @@ from collections import Iterable
 from mpi4py import MPI
 import time
 import random
+import math
 
 
 class ParticleSet(object):
@@ -167,9 +168,9 @@ def recursive_partition(no_proc, dict, sub, dir):
 
     new_dir = 'x'
     cut = -1
-    no_proc_l = math.ceil(no_proc / 2)
-    no_proc_r = math.floor(no_proc / 2)
-    no_part_l = math.ceil(no_proc_l / no_proc * len(sub))
+    no_proc_l = int(math.ceil(no_proc / 2))
+    no_proc_r = int(math.floor(no_proc / 2))
+    no_part_l = int(math.ceil(no_proc_l / no_proc * len(sub)))
 
     if dir == 'x':
         new_dir = 'y'
